# Branchstat algorithm