Refactor of the branch/node stat algorithms. Changes

1. Store the value of f(x) during the branch_stat algorithm, which simplifies the logic a bit and reduces the number of calls of f substantially. (not fully implemented in C when the PR was opened)
2. Abstract out the details of running polarised vs unpolarised summary functions in C for node and branch stats. (It would be pleasing if this was possible for site stats also, but I don't see how as polarisation is much more central to how things are actually computed.)

Should be significant perf gains from this. I'll do some before/after comparisons when its done.