@jeromekelleher 2019-08-29
This can close #257 I think --- getting the function to update now would be very minor perf tweak, probably not worth it. I haven't thought about whether this idea lets us abstract to the AFS yet. That would be nice.