@jeromekelleher 2019-08-29
Benchmarks:

On this tree sequence:
```
$ python3 -m tskit info large-example.trees 
sequence_length:  100000000.0
trees:            456789
samples:          100000
individuals:      0
nodes:            500814
edges:            1941686
sites:            484395
mutations:        484395
migrations:       0
populations:      1
provenances:      1
```
and this code:
```python
import tskit
import time
import numpy as np

ts = tskit.load("large-example.trees")


print("mode", "n_sets", "time", sep="\t")
for mode in ["site", "branch", "node"]:
    for num_sample_sets in [1, 10, 100]:
        sample_sets = np.array_split(ts.samples(), num_sample_sets)
        before = time.perf_counter()
        x = ts.diversity(sample_sets=sample_sets, mode=mode)
        duration = time.perf_counter() - before
        print(mode, num_sample_sets, "{:.3f}".format(duration), sep="\t")
```

On upstream/master we get:
```
$ python3 benchmark.py
mode    n_sets  time
site    1       0.344
site    10      0.557
site    100     2.413
branch  1       1.767
branch  10      2.949
branch  100     15.900
node    1       1.111
node    10      2.162
node    100     12.033
```
On this branch we get 
```
$ python3 benchmark.py
mode    n_sets  time
site    1       0.349
site    10      0.557
site    100     2.426
branch  1       0.647
branch  10      1.761
branch  100     11.555
node    1       0.570
node    10      1.654
node    100     10.541
```

So, we do get a decent perf jump on branch stats when we have a small number of sample sets. Not sure what's going on with the larger number of sample sets, will investigate.