@jeromekelleher 2019-08-29
I don't think we're going to do much better here. When we have 100 sample sets, we get this from perf:
```
  44.93%  python3        _tskit.cpython-36m-x86_64-linux-gnu.so             [.] tsk_treeseq_general_stat                                                                                                                                      
  36.67%  python3        _tskit.cpython-36m-x86_64-linux-gnu.so             [.] diversity_summary_func                                                                                                                                        
  14.79%  python3        _tskit.cpython-36m-x86_64-linux-gnu.so             [.] npolarised_summary_func                                                                                                                                      
   0.65%  python3        [kernel]                                           [k] 0xffffffffb7c009e7                                                                                                                                            
   0.45%  python3        _tskit.cpython-36m-x86_64-linux-gnu.so             [.] tsk_treeseq_init                                                                                                                                              
   0.26%  python3        python3.6                                          [.] _PyEval_EvalFrameDefault           
```
When you dig in we're spending time in looping over the sample sets, as you might expect. We could probably squeeze a bit more perf here by using ``restrict``s, but they are tricky and I don't think we'll get much out of it.

The reason that node and branch stats are so much slower than site stats is because they're doing a lot more. The summary func is only called on mutation nodes not on *every* node. So, it's as good as it's going to get for now I think, and we should be pretty happy with this (it's friggin fast!).