@jeromekelleher 2019-08-29
@petrelharp, any thoughts here? I'm going to merge this and write up the algorithm unless you object.