@petrelharp 2019-09-01
Nice performance improvements!

> Abstract out the details of running polarised vs unpolarised summary functions in C for node and branch stats. (It would be pleasing if this was possible for site stats also, but I don't see how as polarisation is much more central to how things are actually computed.)

This is a nice way to do it. Doing it the same way for site stats would require passing a list of states (one for each derived allele) to a wrapped function, I think - we could work this out but it doesn't seem necessary.